refactor: clean up auth.py and move schemas to auth_schemas.py
